requirements_server_slim2.txt
absl-py==2.1.0
albumentations==1.0.3
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
boto3==1.23.0
botocore==1.26.0
cachetools==5.3.2
certifi==2022.12.7
cffi==1.17.1
channels==4.2.2
chardet==5.2.0
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.0
cryptography==45.0.5
cycler==0.10.0
daphne==4.2.1
distro==1.9.0
dj-database-url==3.0.1
Django==5.0.1
django-allauth==65.10.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
filelock==3.13.1
# fire==0.5.0
fonttools==4.47.2
fsspec==2023.12.2
gevent==25.5.1
gitdb==4.0.11
GitPython==3.1.41
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.20.3
hyperlink==21.0.0
idna==2.10
# imageio==2.33.1
incremental==24.7.2
Jinja2==3.1.3
jiter==0.10.0
jmespath==1.0.1
joblib==1.3.2
jpype1==1.5.2
kiwisolver==1.4.5
konlpy==0.6.0
lazy_loader==0.3
lxml==5.4.0
Markdown==3.5.2
MarkupSafe==2.1.4
# matplotlib==3.8.2
mpmath==1.3.0
mysql-connector-python==9.3.0
# mysqlclient==2.2.1
networkx==3.2.1
# numpy==1.24.3
oauthlib==3.2.2
openai==1.93.0
# opencv-python==4.7.0.72
# opencv-python-headless==4.8.0.74
packaging==23.2
# pandas==2.2.0
# Pillow==10.1.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.7
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybboxes==0.1.6
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyMySQL==1.1.0
pyOpenSSL==25.1.0
pyparsing==2.4.7
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
qudida==0.0.4
requests==2.32.4
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
# roboflow==1.0.0
rsa==4.9
s3transfer==0.5.2
# sahi==0.11.15
# scikit-image==0.22.0
# scikit-learn==1.4.0
# scipy==1.11.4
# seaborn==0.13.2
service-identity==24.2.0
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlparse==0.4.4
sympy==1.12
# tensorboard==2.12.1
# tensorboard-data-server==0.7.2
# tensorboard-plugin-wit==1.8.1
termcolor==2.4.0
terminaltables==3.1.10
# thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.12.9
# torch==2.1.2
# torchvision==0.16.2
tqdm==4.66.1
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2023.4
# ultralytics==8.1.8
uritemplate==4.2.0
urllib3==1.26.6
websocket-client==1.8.0
Werkzeug==3.0.1
# wget==3.2
# yolov5==7.0.13
zope.event==5.1
zope.interface==7.2
